Switch to a different hybrid approach based on exponential step-up. (#1)
This provides a more gradual switch between the linear and binary modes,
rather than making a hard decision at the start based on a guesstimate of
which method will perform better. It also avoids redundant work in the
binary search by focusing the search on the left-most part of the interval.

Nonetheless, it's likely that I'll stick to the linear search as it is simpler and
faster in the majority of cases. Oh well.